+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 6.8.0
+
+ * feat: Added the PSModule AWSPowerShell 
+ * feat: Added the 2nd vmware script to uninstall the tools by removing files and registry entries
+
 ## 6.7.2
 
   * fix: (delete_wave) makes sure the script does not crash even if wave ID is absent